CS-19062 live migration was failed between two pods on advanced isolated network
This is due to a VM on zone wide primary storage not requiring storage migration while migrating across clster. Detecting the storage pool type before allowing normal migration (without storage live migration) of VM across cluster.
listHostsForMigrationOfVM is being called when user attempts to move a VM to other host. This is trying to find list of suitable storage pools that are attached to each of the suitable hosts for the VM. Currently the selection of target storage pools for each volume of the VM is left to storage pool allocators. But user might want to leave his volume unmoved/intact If it is on a zone wide storage pool. This would be more efficient while migrating VM as storage live migration is not required and VM continues to use volumes on same storage pool as before. Hence idea is to set same storage pool as target pool for each of the volume if the volume is already on zone wide storage pool. A comparison of source pool of volume against target pool of volume yields the information if storage migration is required for the VM to move to target host or not. Based on that information apropriate API migrateVM or migrateVmWithVolume could be decided.
Also fixing one more scenario when user directly calls API migrateVirtualMachineWithVolume. If currentPool is accessible to destination host, skip calling allocators and move on to next volume to process. This means if user calls migrateVirtualMachineWithVolume API where all volumes of VM are accessible on specified target host, then API fails as there is no storage migration involved. Instead user should call migrateVirtualMachine API.

Signed-off-by: Sateesh Chodapuneedi <sateesh@apache.org>